// src/resources/CustomItemMap.ts
export enum CustomItemMap {
    // Equipment
    BOREAL_FAST_HELMET = "66ef4fba045adea84e946351",
    BOREAL_FAST_MT_V3 = "67086e83fb45ff5ba709aec5",
    BOREAL_FAST_MT_FINAL = "672a1c74950d97c80032bd5c",
    BOREAL_FAST_SIDE_ARMOR = "66e906e7942ee0fc8b99e4bd",
    BOREAL_FAST_MANDIBLE = "66f0c85c2c885bba5d218566",
    BOREAL_OSPREY_PROTECTION_MTP = "66f63928bf1bd9316a66f3dc",
    BOREAL_OSPREY_ASSAULT = "672a1c4b00492ca9c516a24a",
    BOREAL_COMM3 = "66fc90648cf6d636447517bd",
    BOREAL_REDUT_T5 = "670497be33c79141d703354c",
    BOREAL_D3CRX = "67077dd5451bb1d2b842908a",
    BOREAL_DEATH_WALKER_MASK = "670926092f345e4e2c394eec",
    BOREAL_DEATH_WALKER_MASK_FINAL = "672a1c2354d6662af0fc5463",
    BOREAL_TAKEDOWN = "670c9c93bad875b1c3e21b50",
    BOREAL_RT61_FATMAN = "67768080648a8cc813dd4333",

    // Collectables
    SNAKE_HEAD_CHARM = "6718063b5c8733da95d886b6",

    // Weapon parts & mods
    M4A1_BLUE_RECEIVER = "67171b2e26f750fd69de846c",
    M4A1_BLACK_RECEIVER = "67171bbaf6674de578d8563d",
    GEISSELE_AR15_BLUE = "6716fc5a697b9758fc9774e2",
    GEISSELE_AR15_BLACK = "67171c1d4cb2b2ba3dbff96f",
    AR15_BARREL_508MM = "6716f60a3b058e7df3005054",


    // Clothing
    // Tops
    BOREAL_PARKA_RAGMAN_SUIT = "6731098354bcc7a7281df87c",
    BOREAL_PARKA_SUIT = "6730f9f4e7e28a4639920912",
    BOREAL_PARKA_TOP = "6730fb2deb8293d052ca5bbb",
    BOREAL_PARKA_HANDS = "6730fb3bc3bce93c1ec5000e",

    // Bottoms
    BOREAL_PANTS_BOTTOM = "67315099a2ed0616eab9e07e",
    BOREAL_PANTS_SUIT = "673150ad4cd0ec0c9ba8ef1a",
    BOREAL_PANTS_RAGMAN = "673150c1f6bc9313842e9aa6"
}